# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Sport Energy API
# during dev, should point to your local machine
host: 127.0.0.1:8111
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /sportEnergyAccount/api/v1:
    # binds app logic to a route
    x-swagger-router-controller: sportEnergyAccountController
    get:
      description: Returns sport energy account information to the caller
      # used as the method name of the controller
      operationId: read_sportEnergyAccount
      parameters:
        - in: query
          name: _id
          type: string
          description: The id of the sport energy account
          required: false
        - in: query
          name: cardNumber
          type: string
          description: The card number of the sport energy account
          required: false
      responses:
        200:
          description: Success
    post:
      description: Add sport energy account
      operationId: create_sportEnergyAccount
      parameters:
        - in: body
          name: body
          description: Sport energy account object to be created
          required: true
          schema:
            $ref: '#/definitions/CreateSportEnergyAccount'
      responses:
        200:
          description: Success
    put:
      description: Update sport energy account information
      operationId: update_sportEnergyAccount
      parameters:
        - in: query
          name: _id
          type: string
          description: The id of the sport energy account
          required: true
        - in: body
          name: body
          description: Sport energy account object to be updated
          required: true
          schema:
            $ref: '#/definitions/UpdateSportEnergyAccount'
      responses:
        200:
          description: Success
    delete:
      description: Delete sport energy account
      operationId: delete_sportEnergyAccount
      parameters:
        - in: query
          name: _id
          type: string
          description: The sport energy account to be deleted
          required: true
      responses:
        200:
          description: Success
  /sportEnergyTransaction/api/v1:
    # binds app logic to a route
    x-swagger-router-controller: sportEnergyTransactionController
    get:
      description: Returns sport energy transaction information to the caller
      # used as the method name of the controller
      operationId: read_sportEnergyTransaction
      parameters:
        - in: query
          name: _id
          type: string
          description: The id of the sport energy transaction
          required: false
        - in: query
          name: cardNumber
          type: string
          description: The card number of the sport energy account
          required: false
      responses:
        200:
          description: Success
    post:
      description: Add sport energy transaction
      operationId: create_sportEnergyTransaction
      parameters:
        - in: body
          name: body
          description: Sport energy transaction object to be created
          required: true
          schema:
            $ref: '#/definitions/CreateSportEnergyTransaction'
      responses:
        200:
          description: Success
    delete:
      description: Delete sport energy transaction
      operationId: delete_sportEnergyTransaction
      parameters:
        - in: query
          name: _id
          type: string
          description: The sport energy transaction to be deleted
          required: true
      responses:
        200:
          description: Success
  /sportEnergy/swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  CreateSportEnergyAccount:
    type: object
    required:
    - cardNumber
    - operator
    properties:
      cardNumber:
        type: string
        example: "2090537712048"
      operator:
        type: string
        example: "RFAN27"
    xml:
      name: CreateSportEnergyAccount
  UpdateSportEnergyAccount:
    type: object
    required:
    - operator
    properties:
      cardNumber:
        type: string
        example: "2090537712048"
      energyBalance:
        type: number
        example: 100
      accountStatus:
        type: string
        example: "disabled"
      operator:
        type: string
        example: "RFAN27"
    xml:
      name: UpdateSportEnergyAccount
  CreateSportEnergyTransaction:
    type: object
    required:
    - cardNumber
    - energyUpdate
    - operator
    properties:
      cardNumber:
        type: string
        example: "2090537712048"
      energyUpdate:
        type: number
        example: 300
      operator:
        type: string
        example: "RFAN27"
    xml:
      name: CreateSportEnergyAccount